gpu: Render high depth when node is HDR
When a rendernode is HDR content, we want to render to at least FLOAT16
format, so that the content can reliably be reproduced.

Note that GDK backends may still not support the format, so we might
still end up with a U8 buffer.
In that case, we will continue rendering to U8.