Fix centered text in labels
When the text is rtl, pango will put the text at the right end of
its given width, causing the logical.x to be big, and in turn, our
computed position to be negative. If we don't allow that, centered
text will end up at the right side if it is rtl.

This is only a problem when ellipsizing, since gtk only gives
pango a width in that case.

Fixes: #6836