inspector: Add a button to save a recording as a node video
Take all the recorded nodes and put them into a container node, then
save that container node.

This "video" can then be played back with commands like
  gtk-rendernode-tool show --video
or similar.

Note that this actually achieves a pretty decent compression because
of the deduplication when saving rendernodes and the fact that
unchanging parts of the widget tree will reuse their nodes.