gtkappchooserwidget: Port to GtkListView
This drops the GtkTreeView dependency and the original reason for
deprecation.

Attempted to keep the behavior consistent with the original
implementation:

Sorting - When show_all is FALSE only the Other category is
alphabetically sorted, the other categories keep app info sorting.
When show_all is TRUE, the list is sorted.  One difference is that
in the original implementation, the default app was left as the
first item.  For simplicity (since this was not documented anywhere)
the default app will also be alphabetically sorted, but still
selected by default.

Searching - Type ahead search is maintaining.  It is now
"invisible" (no popover).

Icons - Icons are smaller than in TreeView

Selectability - Now enforces that an item is always selected.
Under certain circumstances the TreeView version would have
no item selected, arguably a bug

All the benefits of list views, proper headers (not fake rows), etc