window: Rework computing minimum size
This function attempts to pick some realistic "minimum size" (as a pair
of width & height) to report to the windowing system. Of course, the GTK
layout machinery doesn't work like that, it has a minimum height for any
given width and the other way around. So this function attempts to pick
one "minimum size" that has a close width/height ratio to the current
size.

The way the binary search was written, it could result in a larger width
or height than the current width or height respectively being reported,
which would then cause the window to unexpectedly enlarge. Make sure
this doesn't happen by not considering widths larger than the current
width, and ones that result in minimum height larger than the current
height.

Rework the logic to:
* never pick a larger width or height than the current values;
* stop if the loop discovers equal heights at both ends;
* avoid throwing away perfectly good potential sizes.

Also rename the variables for clarity.

Helped-by: Mikhail Bugaev <bugaevm.phys@gmail.com>
Co-authored-by: Benjamin Otte <otte@redhat.com>
Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>