boxlayout: Add optimization for the two-child case
In gtk_box_layout_compute_opposite_size (), the box layout performs a
binary search to find the minimum size that would fit its children.
Specifically, distribute_remaining_size () would pick a proposed size
in the opposite orientation, measure every one of the non-constant-size
children for that size along the box's orientation, add their minimum
size requests, and see if it fits the original for-size value; based on
the outcome, it would then consider a larger or a smaller answer.

Note specifically that this always measures the children of the box
along the box's own orientation, even though the box itself is being
measured in the opposite orientation. This can be a good thing or a bad
thing: for instance, if a vertical box containing labels is being
measured width-for-height, the box will binary search for the minimum
width and the labels will be measured height-for-width, which is their
preferred size request mode. If, on the other hand, a horizontal box
containing labels is being measured height-for-width, the box will
binary search for the minimum height, and measure the labels width-for-
height, which is the worst situation possible, since the labels will run
their own binary search to measure themselves that way.

Add a special case code path for the case of a box containing two
variable-size children being measured against the box's own orientation,
but in accordance with their preferred size request mode. In this mode,
the box will instead binary search for the optimal way to split the
available space along the box's orientation among the two children.

For clarity, let's say that we're dealing with a horizontal box
containing two height-for-width children.

It is possible to use binary search this way because of the following
observation:

Suppose we pick some way to split the available width, i.e. some amount
of available width to allocate to the first child, the rest of the width
being allocated to the second child. We then measure the minimum height
of both children for these widths. Suppose the first child requires more
height. Then the only way we could possibly get a better answer (a
smaller minimum height for the box) is by giving more width to the first
child, which may or may not regress the second child's minimum height.
Similarly, if the second child is taller, the only way we could improve
on the result is by giving less space to the first child and more space
to the second child. If both children have the same minimum height, we
can not improve the answer in any way.

It is possible to generalize this algorithm for the case of > 2 child
widgets, but it is significantly more complex, and hits diminishing
returns. Here's the idea of the generalized algorithm anyway:

Between the steps of the search, the state we track is the minimum
amount of width that we're considering allocaitng to each individual
child, i.e. int width_at_least[N]. Of course, the sum of widths cannot
exceed the available width. Initially, the width for each child is equal
to the overall minimum width of the child.

At each step, we consider a way to assign widths to the children.
Specifically, we distribute the remaining (yet undistributed) width
budget among all the children equally (or almost equally if the
remaining width is not a multiple of the number of children), adding it
to the already committed width distribution. We then measure each
child's minimum height for that width, and find out which child requires
most height. The only way we could improve the answer relative to that
point is by giving that child more space. Thus, we commit the amount we
allocated to that child, but not to the others. By doing this, we split
the search space "in half". We then proceeed to the next iteration.

Once there are less units of available width left than the number of
children, we no longer can give *some* additional width to every child
all at the same time, so the final stage of the algorithm will have to
use a different strategy to still make progress: track the height of
each child at the width currently committed to it, and commit one unit
of width to the tallest child, re-measuring it for the new width.

This algorithm reduces the search space by N / (N - 1) times on each
iteration (which is by 2 times for N = 2), so it would need about

log(width, N / (N - 1)) = log2(width) / log2(N / (N - 1))

iterations to converge, with log2(N / (N - 1)) gowing closer to 0 as N
grows, thus the number of iterations growing unboundedly. And each
iteration needs to measure each of the N children once.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>