listbox: Distribute allocated height among children
This requires running looping over the children list several times.
First, we count the number of children and vexpand children, and the
"extra" height. Then we allocate an array of GtkRequestedSize and fill
it with child measurements. We then call
gtk_distribute_natural_allocation () to actually distribute the extra
height among children. Finally, we allocate the children their final
sizes.

We skip a part of this if we find out that we got allocated either or
minimum or natural height.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>