win32: Add GdkD3d12Texture and builder
This is a simple implementation of a GdkTexture that is backed by
a ID3D12Resource.

The only functionality implemented so far is GdkTexture::download(),
but that is enough to make it work - albeit slowly - in all renderers.

It also doesn't support any fancy formats yet; in particular: no YUV.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7883>