win32: Add gdk_d3d12_texture_get_handle()
Makes shared HANDLE handling part of the texture. This handle will be
needed when importing into GL or Vulkan or using elsewhere.
Its lifetime is bound to the texture and so is it's use,
so it's best to maintain it here instead of requiring renderers to
deal with HANDLEs.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7883>