gpu: Add D3D12 texture support to GL renderer
Add a gdk_d3d12_texture_import_gl() function to match the EGL import
behavior.

And then use it in the import code.

What makes this a bit awkward is that we need to not just shuffle a
tex_id but also a mem_id around, but oh well...

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7883>