win32: Replace glsinkbin with d3d12convert on Windows
On Windows, we don't want to use GL and let D3D12 do all the work.

But as long as we don't support all formats and implement all the
interfaces GstPlayer expects, we need to have a way to convert.

So we put a d3d12convert for that.

Related: https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/3956