vulkan: Change the allocator querying API
The old API was awkward in that it didn't allow access to the proper
memory type index and was prone to misuse in codepaths when importing
memory instead of allocating it.

Change that.

The old way now requires more code and a local variable, but importing
memory is a lot more straightforward.

And I don't have to spend days tracking down a leak on Windows.

Part-of: <https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7883>