dmabuf: Port Vulkan to using formats directly
Instead of requiring code in the dmabuf backend to map between fourccs
and VkFormat, query the GdkMemoryFormat and GdkDataFormat enums for their
Vulkan mappings.

This reduces the code duplication quite a lot.

A negative side effect is that unknown formats (to GTK) are no longer
supported. Previously it was possible to map fourccs to VkFormats
without having a corresponding GTK format.