gtk/textview: Use the event's timestamp for (un)obscuring cursors
Where available use the timestamp of the current event to obscure and
unobscure the cursor. Previously we'd only use the timestamp of
the logical pointer but that omitted events from tablet tool, resulting
in the tablet tool's cursor getting hidden on key press but never
re-appearing.

In the case of obscuring triggered by the GtkImContext we resort
to the previous approach of checking the core pointer's timestamp.

Closes #6763