dataformat: Add read_line() vfunc to replace convert()
That function is meant to not do colorspace conversions (ie map from YUV
2 RGB) but just get the pixels unpacked.
It also operates on a single line, so it doesn't have to deal with
strides and it optimizes nicely into multithread conversion.

The disadvatage is that we do the UV conversion per pixel instead of per
block, so this code is slower.

Also added a generic convert() implementation that just calls
read_line().