GdkWin32: Call epoxy_handle_external_wglMakeCurrent
...where we call wglMakeCurrent directly, to inform libepoxy
that the current GL context changed (and thus the current DC
and pixel format may have changed). Skipping the call is not
an issue in practice, since most drivers return the same
pointers regardless of the pixel format (as long as it's an
ICD pixel format instead of the GDI software implementation),
but it's the correct thing to do.