sizerequestcache: Warn on contradictory measurements
When committing the measurement for a widget to the cache, check if the
measurement contradicts a previous measurement in the opposite
orientation, and print a warning. Unfortunately, we don't have a
reference to the widget itself at this point in the code, so we can't
name the specific widget; nevertheless it should be easy to break on the
wanring and inspect the stack trace to see which measurement has
triggered the warning.

This helps debug complex layout issues by catching the contradictory
measurements in the deepest / most nested widget that reports them,
rather than in a some ancestor GtkBox that happens to break because its
children's measurements are contradictory.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>